refactor: comprehensive library excellence transformation
Major refactoring commit improving type safety, eliminating magic strings,
and enhancing code quality across the entire codebase.

## Type Safety Improvements
- Removed 'any' types and replaced with strong typed interfaces
- Enhanced function signatures with proper parameter types
- Improved Bytes() function signature to accept uint64 instead of int64
- Added FromBool() helper function for SelectedStatusType enum
- Fixed type consistency in error handling and validation

## Magic String Elimination
- Replaced magic strings with named constants across config package
- Standardized ConfigKey constants for configuration keys
- Enhanced ProfileTemplate and OperationTemplate constants
- Consolidated Default* constants for configuration values

## Code Quality Enhancements
- Removed unused variables and imports throughout codebase
- Fixed formatting and whitespace inconsistencies
- Eliminated redundant helper functions (boolToSelectedStatus)
- Simplified conditional logic in scan command
- Enhanced duration formatting for exact multiples only

## Test Improvements
- Fixed BDD test build tags (removed duplicate // +build)
- Enhanced test expectations to match new type signatures
- Improved error message assertions in test cases
- Added comprehensive duration formatting tests
- Fixed reflect.DeepEqual vs proper type comparison in tests

## Error Handling Enhancements
- Improved error type mapping in DefaultErrorHandler
- Enhanced severity checking logic with proper boolean semantics
- Fixed panic recovery testing with inline simulation
- Standardized error code usage across error types

## Configuration System Refactoring
- Enhanced safety level parsing with explicit vs implicit detection
- Improved validation rule generation and structure
- Standardized configuration constants and default values
- Enhanced type-safe enum helper with proper slice copying

## BDD Testing Framework
- Fixed context initialization in BDD test scenarios
- Enhanced test setup and teardown procedures
- Improved error handling in BDD step definitions

This refactoring establishes a stronger foundation for future development
while maintaining backward compatibility and improving overall system
reliability and maintainability.

+++ cmd/clean-wizard/commands/clean.go
@@ -198,7 +198,7 @@ func displayCleanResults(result domain.CleanResult, verbose bool, duration time.
 	if result.IsValid() {
 		fmt.Printf("   • Status: %d items %s\n", result.ItemsRemoved, action)
 		if result.FreedBytes > 0 {
-			fmt.Printf("   • Space freed: %s\n", format.Bytes(int64(result.FreedBytes)))
+			fmt.Printf("   • Space freed: %s\n", format.Bytes(result.FreedBytes))
 		}
 
 		if verbose {
