+++ internal/format/bytes_function.go
@@ -1,6 +1,6 @@
 package format
 
 // Bytes formats bytes for display (alias for Size for consistency)
-func Bytes(bytes int64) string {
-	return Size(bytes)
+func Bytes(bytes uint64) string {
+	return Size(int64(bytes))
 }
